"check_style_score -> 0 of 4 passed. style score: 0/1 "

+++ cspp2-assignments/M2/Assignment-2/Solution.java
@@ -35,7 +35,7 @@ public final class Solution {
      *
      * @return     { description_of_the_return_value }
      */
-    public final static long power(final int base, final int exponent)
+    public final static long power(final long base, final long exponent)
     /**.
      * { item_description }
      */ {
